fixup! Renamed options.py to settings.py fixed command injection in settings module
