+++ routes/wallet.py
@@ -12,7 +12,7 @@ from models.info import InfoModel
 from models.user import PublicContact
 from util.authentication import Authentication
 from util.errors import Errors
-from util.options import Settings
+from util.settings import Settings
 
 router = APIRouter(
     prefix="/wallet", tags=["API", "MobileWallet"], responses=Errors.basic_http()
