+++ util/email.py
@@ -4,7 +4,7 @@ from email.mime.text import MIMEText
 
 import commonmark
 
-from util.options import Settings
+from util.settings import Settings
 
 email = Settings().email.email
 password = Settings().email.password.get_secret_value()
